test: Add more delay in txs on upgrade test

+++ integration_test/upgrade/scripts/kiichaind_upgrade_bin.sh
@@ -11,7 +11,13 @@ export GOBIN=$GOPATH/bin
 
 # kill the existing service
 pkill -f "kiichaind start"
-sleep 5
+sleep 10
+
+# Verify the process is no longer running
+if pgrep -f "kiichaind start"; then
+  echo "Error: kiichaind process is still running."
+  exit 1
+fi
 
 # Replace the binary
 cp build/kiichaind "$GOBIN"/
@@ -20,7 +26,7 @@ cp build/kiichaind "$GOBIN"/
 kiichaind start --chain-id kii --inv-check-period ${INVARIANT_CHECK_INTERVAL} > "$LOG_DIR/kiichaind-$NODE_ID.log" 2>&1 &
 
 # Sleep to catch-up consensus
-sleep 5
+sleep 30
 
 echo "PASS"
 exit 0
